requirements.txt
streamlit
torch
onnxruntime
Pillow
pandas
numpy
plotly
groq
youtube-search
python-dotenv